Adding support to DX11 to create a WARP Device
This is not enabled by default, and will emit a line of log if it is enabled.

It can be enabled by setting: trinity.device.deviceType = trinity.TriDeviceType.SOFTWARE
or through other methods that affect the presentation parameters.

This is useful for automated testing without a 3D device, although it requires DX10 or 11, and
is only really fully supported on DX11.1 (Win8 / Server 2012 R2).

CR Filipp

Binaries submitted in 808681
Build time : 9 minutes 42.445000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/": change = 808630]