Fighters cannot be manipulated or selected within jessica once loaded https://jira.ccpgames.com/browse/EVEDEV-1189
Added the ITr2Pickable interface to EveSwarmRenderable, fixed the bounding box which behaved a bit strangely.
Ended up fixing shadows for fighters as when the bounding box was fixed it ended up in a shadow being cast as if the was a static fighter at the center of the bounding box.

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1123748]