SOF bannersets
Now I had to do somethings that might look strange because of how it was initially written:
The old way of getting banners onto spaceobjects we have banners on the hull which is then converted into HullBannerSetData and then added to the space object attachment.
I converted that to be like this: banner on hull -> HullBannerData -> add to space object attachment.

This is so I could have a clearer way of getting bannersets down to the space object attachment:
bannnerset on hull -> HullBannerSetData -> space object attachment


https://jira.ccpgames.com/browse/EO-5082

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1842938]