Blue exposed tonemapping and new lighting variables
Added blue exposed tonemapping effect for easy authoring.
If no tonemapping is configured then we use eve tonemapping.

Added a way to make the cubemap brighter in the reflection map.

Also added a way to configure the lowest mip for diffuse nebula reflection. It is in roughness space (1 is most rough therefore the lowest resolution mip and 0 is most smooth and therefore the highest resolution mip)

https://ccpgames.atlassian.net/browse/PLAT-5709
https://ccpgames.atlassian.net/browse/PLAT-5713