Upscaling lod fixes
Removing all usage of global lod variables.
We now push this into the EveUpdateContext.
This meant that a lot of things now get a const EveUpdateContext& instead of a const TriFrustum& in the UpdateVisibility functions.
I also changed all EveUpdateContext& parameters to const EveUpdateContext&

I ALSO had to fix a few files that were using the global variables in GetRenderables and some other random functions.

https://ccpgames.atlassian.net/browse/PLAT-5931