Fixing device lost
Added slFreeResources to the Tr2StreamlineAL which is needed when we destroy / reuse a dlss context, because Streamline holds on to ... something ...

Removed the use of NIS (Nvidia Image Sharpener) because it has a bug in it (it will never remove its viewports from the nis context somewhere deep in streamline), I had fixed the bug before by always setting the viewhandle to 0, but I found another edge case when you fire up the planetary interaction (then we have 2 contexts).

Also removed the use of setTagsForFrame since we just call evaluate in the same code path, so sending the resources that way is hopefully faster.

Added a struct that keeps textures around until the next time they are set (ResourceReference) or deleted.

Tagged the resources as validuntilevaluate, except the hudless texture which is only valid now (just to be safe)